requirements.txt
boto3
awslambdaric
pg8000
python-dotenv